Resolved pull request issues
