+++ src/Components/bookHub/BookHub.js
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from 'react';
-import RangeFilter from '../Filtering/RangeFilter';
+import Filter from '../Filtering/Filtering';
 import Card from './BookCard';
 
 const gradeRangeMetadata = ['0 to Pre-K', '1st', '2nd', '3rd', '4th', '5th', '6th', '7th', '8th', '9th', '10th', '11th', '12th'];
 const ageRangeMetadata = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12', '13', '14', '15', '16', '17', '18'];
-// airtable configuration
+
 const Airtable = require('airtable');
 
 const airtableConfig = {
@@ -77,12 +77,11 @@ function CardsDisplay() {
 
   useEffect(() => { getCards(); }, []);
   useEffect(() => {
-    console.log(filteredCards);
   }, [filteredCards]);
 
   return (
     <div>
-      <RangeFilter
+      <Filter
         setRangeState={setRangeInput}
         setMultiSelectInput={setMultiSelectInput}
       />
